requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
python-dotenv
gunicorn
whitenoise
